@@ -17,10 +17,10 @@ public class RenderingTests extends BaseLDBCTests {
     @Test
     @Disabled
     public void testBasicCypherToSQL() throws Exception {
-        String cypher = "" +
-                "MATCH (a:Person {firstName: 'Richard', lastName: 'Smith'})-[:KNOWS]-(b:Person)\n" +
-                "MATCH (a {id: 0})-[:KNOWS]-(c:Person)-[:KNOWS]-(b)\n" +
-                "RETURN b.firstName";
+        String cypher = """
+                MATCH (a:Person {firstName: 'Richard', lastName: 'Smith'})-[:KNOWS]-(b:Person)
+                MATCH (a {id: 0})-[:KNOWS]-(c:Person)-[:KNOWS]-(b)
+                RETURN b.firstName""";
         cypher2sql(cypher);
     }
 
@@ -34,20 +34,20 @@ public class RenderingTests extends BaseLDBCTests {
 
     @Test
     public void testCountFunction() throws Exception {
-        String cypher = "" +
-                "MATCH (a:Message)\n" +
-                "WITH count(a) AS messageCount\n" +
-                "RETURN messageCount";
+        String cypher = """
+                MATCH (a:Message)
+                WITH count(a) AS messageCount
+                RETURN messageCount""";
         cypher2sql(cypher);
     }
 
     @Test
     public void testAliasVarInWith() throws Exception {
-        String cypher = "" +
-                "MATCH (a:Person)\n" +
-                "WITH a AS b\n" +
-                "MATCH (b)-[:KNOWS]-(c:Person) WHERE c.firstName = b.firstName\n" +
-                "RETURN b.firstName, c.age";
+        String cypher = """
+                MATCH (a:Person)
+                WITH a AS b
+                MATCH (b)-[:KNOWS]-(c:Person) WHERE c.firstName = b.firstName
+                RETURN b.firstName, c.age""";
         cypher2sql(cypher);
     }
 
@@ -55,19 +55,19 @@ public class RenderingTests extends BaseLDBCTests {
     public void testDisjointMatch() throws Exception {
         // This is example should trigger the second MATCH
         // to increase the cardinality of the final result.
-        String cypher = "" +
-                "MATCH (a:Person {id:0})-[:KNOWS]-(b:Person)\n" +
-                "MATCH (c:Person)\n" +
-                "RETURN 42";
+        String cypher = """
+                MATCH (a:Person {id:0})-[:KNOWS]-(b:Person)
+                MATCH (c:Person)
+                RETURN 42""";
         cypher2sql(cypher);
     }
 
     @Test
     public void testMultipleMatchOnNode() throws Exception {
-        String cypher = "" +
-                "MATCH (a:Message)\n" +
-                "RETURN a.length ORDER BY a.length DESC\n" +
-                "LIMIT 10";
+        String cypher = """
+                MATCH (a:Message)
+                RETURN a.length ORDER BY a.length DESC
+                LIMIT 10""";
         cypher2sql(cypher);
     }
 
@@ -79,57 +79,57 @@ public class RenderingTests extends BaseLDBCTests {
 
     @Test
     public void testMergeExpandCyphers() throws Exception {
-        String cypher = "" +
-                "MATCH (a)\n" +
-                "RETURN a.id\n" +
-                "LIMIT 10";
+        String cypher = """
+                MATCH (a)
+                RETURN a.id
+                LIMIT 10""";
         System.out.println(PassAvailables.cypher2sql(getGTopInterface(), cypher));
     }
 
     @Test
     public void testIndirectDependency() throws Exception {
-        String cypher = "" +
-                "MATCH (a:Person)-[:KNOWS]-(b:Person)\n" +
-                "MATCH (b)-[:KNOWS]-(c:Person)\n" +
-                "MATCH (a)-[:KNOWS]-(d:Person)\n" +
-                "RETURN 42";
+        String cypher = """
+                MATCH (a:Person)-[:KNOWS]-(b:Person)
+                MATCH (b)-[:KNOWS]-(c:Person)
+                MATCH (a)-[:KNOWS]-(d:Person)
+                RETURN 42""";
         cypher2sql(cypher);
     }
 
     @Test
     public void testFullExample() throws Exception {
-        String cypher = "" +
-                "MATCH (test:Person {id:2199023259437})-[:KNOWS]->(friend:Person)\n" +
-                "MATCH (:Person)-[:KNOWS]->(other_friend:Person)\n" +
-                "MATCH (friend)-[:STUDY_AT]->(:University)-[:IS_LOCATED_IN]->(:City)\n" +
-                "RETURN friend.id";
+        String cypher = """
+                MATCH (test:Person {id:2199023259437})-[:KNOWS]->(friend:Person)
+                MATCH (:Person)-[:KNOWS]->(other_friend:Person)
+                MATCH (friend)-[:STUDY_AT]->(:University)-[:IS_LOCATED_IN]->(:City)
+                RETURN friend.id""";
         cypher2sql(cypher);
     }
 
     @Test
     public void testCircularRelationships() throws Exception {
-        String cypher = "" +
-                "MATCH (a:Person)-[:KNOWS]-(b:Person)-[:KNOWS]-(c:Person)\n" +
-                "MATCH (a)-[:KNOWS]-(c)\n" +
-                "RETURN a.id, b.id, c.id";
+        String cypher = """
+                MATCH (a:Person)-[:KNOWS]-(b:Person)-[:KNOWS]-(c:Person)
+                MATCH (a)-[:KNOWS]-(c)
+                RETURN a.id, b.id, c.id""";
         cypher2sql(cypher);
     }
 
     @Test
     public void testJoinsExpandedIntoProblematicUnions() throws Exception {
-        String cypher = "" +
-                "MATCH (a:Message)\n" +
-                "MATCH (a)<-[:LIKES]-(b:Person)\n" +
-                "RETURN a.id, b.id";
+        String cypher = """
+                MATCH (a:Message)
+                MATCH (a)<-[:LIKES]-(b:Person)
+                RETURN a.id, b.id""";
         cypher2sql(cypher);
     }
 
     @Test
     public void testAliasVar() throws Exception {
-        String cypher = "" +
-                "MATCH (a:Person)-[:KNOWS]-(b:Person)\n" +
-                "WITH a, {c:{b:b}} AS d\n" +
-                "RETURN a.firstName, d.c.b.firstName";
+        String cypher = """
+                MATCH (a:Person)-[:KNOWS]-(b:Person)
+                WITH a, {c:{b:b}} AS d
+                RETURN a.firstName, d.c.b.firstName""";
         cypher2sql(cypher);
     }
 
@@ -143,10 +143,10 @@ public class RenderingTests extends BaseLDBCTests {
 
     @Test
     public void testAliasVarInUnions() throws Exception {
-        String cypher = "" +
-                "MATCH (a:Message)\n" +
-                "WITH a, {c:{b:a}} AS d\n" +
-                "RETURN a.firstName, d.c.b.firstName";
+        String cypher = """
+                MATCH (a:Message)
+                WITH a, {c:{b:a}} AS d
+                RETURN a.firstName, d.c.b.firstName""";
         cypher2sql(cypher);
     }
 
